requirements.txt
flask==3.0.0
gunicorn==21.2.0
google-generative-ai==0.3.1
